Fix bug #25 off-by-one in create-eboot string table builder
Fixed an issue reported by IDC where the string table builder in the .sce_dynlib_data section was recording a size that was 1 byte under what it should have been for the DT_SCE_STRSZ dynamic table entry. This was due to a bug where the size of the string table was not always incremented by 1 to account for the leading null byte depending on how the padding was calculated due to a nasty logic bug.